+++ NEWS.md
@@ -4,6 +4,7 @@
 * Add support for the time course automatic interval option.
 * Fix definition of experimental data with columns of type 'ignore'.
 * Fix `runParameterEstimation()` output, in some cases, listing incorrect parameter start values.
+* Minor fixes and code cleanups.
 
 # CoRC 0.7.1 (2020-05-24)